feat(buzzer): kind 0 = remote stop, kind >= 16 collapses to 0, shell repeat_s (#338)
Review follow-ups agreed with the maintainer:

- kind 0 over buzzer_play is now the STOP request — the remote counterpart
  of `ats buzzer off` (release builds have no shell, so this was the only
  missing way to stop a repeating melody over NFC/LRW). It purges a queued
  request, aborts a mid-playback beep/melody, and cancels a pending repeat
  cycle. An empty BuzzerPlay (proto3 default) therefore also reads as stop.
- Any kind >= 16 collapses to 0 — the wire id is a nibble; wider values no
  longer alias into the 1-15 range via uint8 truncation (the API now takes
  uint32 end-to-end and app_buzzer.c does the collapsing).
- `ats buzzer off` routes through the same kind-0 stop path (purge + abort)
  instead of a bare pin write, so local and remote stop behave identically.
- `ats buzzer play <melody> [repeat_s]` gained the optional repeat argument
  so the repeat loop is bench-testable without a real NFC/LRW downlink.
- Doc: buzzer_play's Ack documented as fire-and-forget (confirms acceptance,
  not audible playback; newest request supersedes a not-yet-started one).

Verified: pytest configen (26/26), tests/cmd suite (29/29, kind-0 forwarding
case added), clang-format clean, release 70.81%/78.73% + debug 99.36%/99.14%
builds pass.

+++ app/src/app_cmd.c
@@ -791,11 +791,12 @@ static void app_cmd_handle_clm_rearm(enum app_cmd_transport tp, const Command *c
 
 /* #338: remote-triggered buzzer melody (NFC/LRW). kind selects one of the
  * fixed severity melodies (the buzzer is DC self-oscillating only — no custom
- * tones); repeat_s (0-999s), if non-zero, keeps re-playing the melody that
- * many seconds after each playback finishes, until stopped locally
- * (`ats buzzer off`, which also aborts a beep/melody mid-playback rather than
- * waiting it out) — there's no protocol-level "stop", by design: this is a
- * fire-and-forget trigger, not a persisted alarm state on the wire. */
+ * tones); kind 0 (or any id >= 16) is the STOP request — the remote
+ * counterpart of `ats buzzer off`, silencing a beep/melody mid-playback and
+ * cancelling a pending repeat cycle. repeat_s (0-999s), if non-zero, keeps
+ * re-playing the melody that many seconds after each playback finishes, until
+ * stopped. Fire-and-forget: the Ack confirms the request was accepted, not
+ * that it audibly played (a newer request supersedes an older queued one). */
 static void app_cmd_handle_buzzer_play(enum app_cmd_transport tp, const Command *cmd,
 				       Response *resp, enum app_cmd_action *action)
 {
@@ -808,11 +809,11 @@ static void app_cmd_handle_buzzer_play(enum app_cmd_transport tp, const Command 
 		return;
 	}
 
-	/* kind is a raw id (0-15, #338) — this dispatch layer never validates or
+	/* kind is a raw id (#338) — this dispatch layer never validates or
 	 * interprets it; app_buzzer_play_repeating() (app_buzzer.c MELODY_TABLE)
-	 * is the sole authority on which ids exist, so a new melody never needs a
-	 * change here. */
-	int ret = app_buzzer_play_repeating((uint8_t)play->kind, (uint16_t)play->repeat_s);
+	 * is the sole authority on which ids exist (0 / >=16 = stop, 1-3 =
+	 * melodies, 4-15 reserved), so a new melody never needs a change here. */
+	int ret = app_buzzer_play_repeating(play->kind, (uint16_t)play->repeat_s);
 	if (ret == -ENOENT) {
 		make_error(resp, Response_Error_Code_BAD_REQUEST, "unknown melody kind");
 		return;
